Recover ImperialismApp's missing message-map handlers (#402)
* Recover ImperialismApp's missing message-map handlers

Six more of the 267d66d8d stub wave, all named IltTarget_*. Finding their owner
took a detour worth recording: they have no xrefs in the Ghidra DB at all,
because the ILT thunks that reach them were pruned, and the thunk JMP edges went
with them. Scanning the original PE on disk instead (absolute dwords + E8/E9
rel32) found each one behind exactly one ILT thunk.

Two of those thunks are referenced from .rdata below every known vtable. That
region decodes as 24-byte MFC AFX_MSGMAP_ENTRY records:

  nMessage=0x111 (WM_COMMAND) | nCode=-1 | nID=0x801e | nLastID=0x801e |
  nSig=0x2c (AfxSig_vpv) | pfn=0x00407e73 -> 0x004147d0

Walking that array back to its AFX_MSGMAP header and forward to its terminator
gives 17 entries, referenced from 0x412aa1 -- the body of
ImperialismApp::GetMessageMap. So these are ImperialismApp command handlers, and
our BEGIN_MESSAGE_MAP had only 6 of the original's 17 entries.

Added the six that have a recoverable owner, in the original's slot order:

  0x00413d00  OnDeveloperCommand8014        ON_COMMAND(0x8014)
  0x004147b0  OnDeveloperCommand801E        ON_COMMAND(0x801e)
  0x004147d0  OnUpdateDeveloperCommand801E  ON_UPDATE_COMMAND_UI(0x801e)
  0x004147f0  OnDeveloperCommand801F        ON_COMMAND(0x801f)
  0x00414810  OnUpdateDeveloperCommand801F  ON_UPDATE_COMMAND_UI(0x801f)
  0x00414830  OnDeveloperCommand8020        ON_COMMAND(0x8020)

plus the two standard CWinApp entries the map really carries (0xe100 ID_FILE_NEW
and 0xe101 ID_FILE_OPEN). All six compare at 100%.

0x8014 tail-calls the retail no-op module-library cache hook
(mov ecx,[g_pModuleLibraryCacheState]; jmp NoOpRetailCacheHook). The other five
are empty in retail -- the developer features behind them were compiled out,
leaving only the message-map wiring -- so they are named after their command ID
rather than given invented semantics (Hard Rule 6).

Not added, because they have no clean owner yet and would need their own
recovery: 0xe140 ID_APP_ABOUT -> 0x413700, 0x801d -> 0x414790 (currently inside
another function's extent), and 0x64 -> 0x413930 (inside IltTarget_004138e0).

Also fixes a real pre-existing type defect this surfaced.
ComputeTileIndexFromHexColumnX2AndRow (0x512850) declared `int columnX2`, but
the retail listing opens `movsx eax, word ptr [esp+4]` -- the parameter is a
short. The body was already right; only the type was wrong. Narrowing it takes
the function 55.56% -> 100%. It had dipped to 31.58% here first: growing this
binary reshuffled its register schedule, which is what made the long-standing
mismatch visible.

Verified: just precommit exit 0. average similarity 77.74% -> 77.80% (+0.06pp),
exact/original 51.13% -> 51.22% (+0.09pp), 10 improved. Two residual dips,
TMultiplayerMgr::ProcessDiplomacyTurnStateEventStateMachine (-1.09pp) and
BuildHexAreaTileIndexList (-0.71pp), are untouched code; just triage returns
inconclusive for both, which is not evidence of a source defect, so the source
is left alone.


* Refresh reccmp baseline after main merge


---------

+++ src/game/map/TMapMgr.cpp
@@ -1665,7 +1665,7 @@ void SplitTileIndexToHexRasterColumnX2AndRow(StrategicTileIndex tileIndex, short
 // Combines a doubled hex-raster column (columnX2, as produced by
 // SplitTileIndexToHexRasterColumnX2AndRow) and a row back into a linear tile index.
 // FUNCTION: IMPERIALISM 0x00512850
-int ComputeTileIndexFromHexColumnX2AndRow(int columnX2, int row) {
+int ComputeTileIndexFromHexColumnX2AndRow(short columnX2, int row) {
   return columnX2 / 2 + row * 0x6c;
 }
